requirements.txt
# MBWCT-Net 项目依赖
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
scipy>=1.7.0
tqdm>=4.62.0
seaborn>=0.11.0
tensorboard>=2.6.0
pytorch-wavelets==1.3.0